+++ tests/api_contract.rs
@@ -42,7 +42,7 @@ fn context(base_url: String) -> AppContext {
             worker_id: Some("00000000-0000-4000-8000-000000000001".into()),
             tenant_id: Some("00000000-0000-4000-8000-000000000002".into()),
             worker_name: Some("contract-worker".into()),
-            credential_store: Some("environment".into()),
+            credential_store: Some("private_file_fallback".into()),
             credential_expires_at_unix: None,
             project_id: None,
             project_key: None,
@@ -75,7 +75,7 @@ fn context(base_url: String) -> AppContext {
         tenant_id: Some("00000000-0000-4000-8000-000000000002".into()),
         worker_name: Some("contract-worker".into()),
         installation_id: "00000000-0000-4000-8000-000000000099".into(),
-        credential_source: rustgrid_agent::credentials::CredentialSource::Environment,
+        credential_source: rustgrid_agent::credentials::CredentialSource::FallbackFile,
         credential_expires_at_unix: None,
         credential_store: rustgrid_agent::credentials::CredentialStore::new(
             &base_url,